+++ R/callbacks.R
@@ -207,7 +207,10 @@ callback_terminate_on_naan <- function() {
 #'   [details](https://www.tensorflow.org/how_tos/embedding_viz/#metadata_optional)
 #'    about the metadata file format. In case if the same metadata file is used
 #'   for all embedding layers, string can be passed.
-
+#' @param embeddings_data Data to be embedded at layers specified in
+#'   `embeddings_layer_names`. Array (if the model has a single input) or list 
+#'   of arrays (if the model has multiple inputs). Learn [more about embeddings](https://www.tensorflow.org/programmers_guide/embedding)
+#'
 #' @details TensorBoard is a visualization tool provided with TensorFlow.
 #'   
 #' You can find more information about TensorBoard
@@ -227,7 +230,8 @@ callback_tensorboard <- function(log_dir = NULL, histogram_freq = 0,
                                  write_images = FALSE,
                                  embeddings_freq = 0, 
                                  embeddings_layer_names = NULL,
-                                 embeddings_metadata = NULL) {
+                                 embeddings_metadata = NULL,
+                                 embeddings_data = NULL) {
   
   # establish the log_dir
   if (is.null(log_dir)) {
@@ -244,12 +248,16 @@ callback_tensorboard <- function(log_dir = NULL, histogram_freq = 0,
     write_images = write_images
   )
   
+  if (!missing(embeddings_data) && keras_version() < "2.2.0")
+    stop("embeddings_data requires keras >= 2.2. Please update with install_keras()")
+  
   # embeddings arguments seem to have been excluded in the TF implementation
   # (even though they are stil part of the docs there)
   if (!is_tensorflow_implementation()) {
     args$embeddings_freq <- as.integer(embeddings_freq)
     args$embeddings_layer_names <- embeddings_layer_names
     args$embeddings_metadata <- embeddings_metadata
+    args$embeddings_data <- embeddings_data
   }
   
   if (keras_version() >= "2.0.5") {