+++ inst/python/kerastools/model.py
@@ -3,7 +3,10 @@
 import os
 
 if (os.getenv('KERAS_IMPLEMENTATION', 'keras') == 'tensorflow'):
-  from tensorflow.python.keras.engine import Model
+  try:
+    from tensorflow.python.keras.engine import Model
+  except:
+    from tensorflow.python.keras.engine.training import Model
 else:
   from keras.engine import Model
  
@@ -13,5 +16,5 @@ class RModel(Model):
   def __init__(self, name = None):
     super(RModel, self).__init__(name = name)
  
-  def call(self, inputs, mask = None):
-    return self._r_call(inputs, mask)
+  def call(self, inputs, mask = None, **kwargs):
+    return self._r_call(inputs, mask, **kwargs)