+++ NEWS.md
@@ -39,6 +39,8 @@ Install the development version with: `install_github("rstudio/keras")`
 
 - New `custom_metric()` function for defining custom metrics in R
 
+- Provide typed wrapper for categorical custom metrics
+
 - Provide access to Python layer within R custom layers
 
 - Don't convert custom layer output shape to tuple when shape is a list